Remove test project and related files
The `InventoryManagmentTests.csproj` and its associated
test files have been completely removed, including all
references to testing frameworks like `xunit`, `Moq`,
and `FluentAssertions`. The solution file has been updated
to reflect the removal of the test project from the
solution.

This needs to be done in order for the new implementation and refactor for the new interface.